refactor(daemon): cc_gui_daemon.rs 七个内联 mod 落盘
- 3076→2555 行；state/event_sink/remote_backend/session_index/shared_sessions/codex/files
  落为 bin/cc_gui_daemon/ 同名文件（#[path] 先例）
- include_str! 自引用回归测试断言仍命中（2 passed）
- baseline 重生成 + 新文件登记豁免
验证：cargo check --bin cc_gui_daemon 零 error；cargo test --bin cc_gui_daemon
失败集与基线完全一致（9 个既有环境失败，与本次无关）
