+++ docs/architecture/large-file-baseline.json
@@ -1,5 +1,5 @@
 {
-  "generatedAt": "2026-08-30T15:21:23.931Z",
+  "generatedAt": "2026-08-30T15:35:38.954Z",
   "scope": "fail",
   "policyVersion": "2026-08-30.policy-v5",
   "entries": [
@@ -371,16 +371,6 @@
       "priority": "P1",
       "warnThreshold": 2200,
       "failThreshold": 2800,
-      "thresholdSource": "policy"
-    },
-    {
-      "path": "src-tauri/src/bin/cc_gui_daemon.rs",
-      "lines": 3076,
-      "type": "rust",
-      "policyId": "default-source",
-      "priority": "P1",
-      "warnThreshold": 2600,
-      "failThreshold": 3000,
       "thresholdSource": "policy"
     },
     {
