+++ src-tauri/src/bin/cc_gui_daemon.rs
@@ -52,74 +52,12 @@ mod rpc_params;
 mod snapshot_throttle;
 #[path = "../shared_binding_visibility.rs"]
 mod shared_binding_visibility;
-// `local_usage.rs` is shared with the desktop Tauri app and references
-// `crate::state::AppState` in command wrappers. The daemon only reuses the
-// workspace-backed filesystem helpers, so a minimal stub keeps the shared
-// module compilable here without pulling the full desktop app state graph.
-mod state {
-    use std::collections::HashMap;
-    use std::sync::Arc;
-    use tokio::sync::Mutex;
-
-    use crate::backend::app_server::WorkspaceSession;
-    use crate::engine::EngineManager;
-    use crate::runtime::RuntimeManager;
-    use crate::types::{AppSettings, WorkspaceEntry};
-    use std::path::PathBuf;
-
-    #[allow(dead_code)]
-    pub(crate) struct AppState {
-        pub(crate) workspaces: Mutex<HashMap<String, WorkspaceEntry>>,
-        pub(crate) sessions: Mutex<HashMap<String, Arc<WorkspaceSession>>>,
-        pub(crate) app_settings: Mutex<AppSettings>,
-        pub(crate) storage_path: PathBuf,
-        pub(crate) settings_path: PathBuf,
-        pub(crate) runtime_manager: Arc<RuntimeManager>,
-        pub(crate) engine_manager: Arc<EngineManager>,
-    }
-
-    impl AppState {
-        #[allow(dead_code)]
-        pub(crate) async fn sync_engine_configs_from_settings(&self) {}
-    }
-}
-mod event_sink {
-    use tauri::AppHandle;
-
-    use crate::backend::events::{AppServerEvent, EventSink, TerminalOutput};
-
-    #[derive(Clone)]
-    pub(crate) struct NoopEventSink;
-
-    impl EventSink for NoopEventSink {
-        fn emit_app_server_event(&self, _event: AppServerEvent) {}
-
-        fn emit_terminal_output(&self, _event: TerminalOutput) {}
-    }
-
-    pub(crate) fn build_event_sink(_app: AppHandle) -> NoopEventSink {
-        NoopEventSink
-    }
-}
-mod remote_backend {
-    use serde_json::Value;
-    use tauri::AppHandle;
-
-    use crate::state::AppState;
-
-    pub(crate) async fn is_remote_mode(_state: &AppState) -> bool {
-        false
-    }
-
-    pub(crate) async fn call_remote(
-        _state: &AppState,
-        _app: AppHandle,
-        _method: &str,
-        _params: Value,
-    ) -> Result<Value, String> {
-        Err("remote backend is not available in daemon".to_string())
-    }
-}
+#[path = "cc_gui_daemon/state.rs"]
+mod state;
+#[path = "cc_gui_daemon/event_sink.rs"]
+mod event_sink;
+#[path = "cc_gui_daemon/remote_backend.rs"]
+mod remote_backend;
 #[allow(dead_code)]
 #[path = "../agent_catalog.rs"]
 mod agent_catalog;
@@ -134,319 +72,16 @@ mod rules;
 #[allow(dead_code)]
 #[path = "../runtime/mod.rs"]
 mod runtime;
-// session_management.rs 的删除核心在收口时会打 session index tombstone；
-// daemon 没有本地 SQLite index（web 模式走 legacy list），用 no-op 保持
-// 共享核心可编译，行为与桌面端一致（桌面端走完整 session_index 模块）。
-#[allow(dead_code)]
-mod session_index {
-    pub(crate) mod commands {
-        pub(crate) async fn tombstone_session_index_rows(
-            _session_ids: Vec<String>,
-        ) -> Result<u32, String> {
-            Ok(0)
-        }
-    }
-
-    // session_management overlay types; daemon has no SQLite index.
-    pub(crate) mod store {
-        #[derive(Debug, Clone, Default)]
-        pub struct SessionIndexRow {
-            pub engine: String,
-            pub session_id: String,
-            pub title: String,
-            pub native_title: Option<String>,
-            pub updated_at: i64,
-            pub created_at: Option<i64>,
-            pub cwd: Option<String>,
-            pub workspace_path: Option<String>,
-            pub physical_path: Option<String>,
-            pub parent_session_id: Option<String>,
-            pub size_bytes: Option<u64>,
-            pub provider_profile_id: Option<String>,
-            pub provider_profile_name: Option<String>,
-        }
-
-        pub(crate) const INDEX_LIST_ENGINES: &[&str] = &[
-            "claude", "codex", "gemini", "grok", "kimi", "opencode", "pi", "dsh", "qoder",
-        ];
-
-        #[derive(Debug, Clone)]
-        pub(crate) struct SessionIndexDeleteLookup {
-            pub(crate) row: SessionIndexRow,
-            pub(crate) tombstoned_at: Option<i64>,
-        }
-
-        /// daemon 无本地 SQLite index：open 恒失败，archive v2 resolve 走
-        /// engine 前缀定向 + 请求 workspace 回退（与桌面端 index miss 同语义）。
-        pub(crate) struct DaemonUnavailableConnection;
-
-        pub(crate) fn open_connection() -> Result<DaemonUnavailableConnection, String> {
-            Err("daemon has no local session index".to_string())
-        }
-
-        pub(crate) fn lookup_rows_for_delete(
-            _connection: &DaemonUnavailableConnection,
-            _full_id: &str,
-        ) -> Result<Vec<SessionIndexDeleteLookup>, String> {
-            Ok(Vec::new())
-        }
-    }
-
-    // tombstone_filter no-op：daemon 无本地 index，过滤器恒空（与桌面端
-    // fail-open 语义一致，共享的 session_management 出口过滤编译可用）。
-    pub(crate) mod tombstone_filter {
-        #[derive(Debug, Default)]
-        pub(crate) struct TombstoneFilter;
-
-        impl TombstoneFilter {
-            pub(crate) fn load_fail_open() -> Self {
-                Self
-            }
-
-            pub(crate) fn is_empty(&self) -> bool {
-                true
-            }
-
-            pub(crate) fn is_tombstoned(&self, _engine: &str, _session_id: &str) -> bool {
-                false
-            }
-
-            #[allow(dead_code)]
-            pub(crate) fn retain<T>(
-                &self,
-                _engine: &str,
-                _sessions: &mut Vec<T>,
-                _id_of: impl Fn(&T) -> &str,
-            ) {
-            }
-        }
-    }
-}
+#[path = "cc_gui_daemon/session_index.rs"]
+mod session_index;
 
 #[allow(dead_code)]
 #[path = "../session_archive_v2.rs"]
 mod session_archive_v2;
 // session_management now catalogs/deletes shared sessions via crate::shared_sessions.
 // The desktop app gets the full module from lib.rs; the daemon keeps a small local
-// adapter and only uses the shared read-only V2 binding projection. It never pulls
-// SharedEventWriter or the full Tauri command graph into the daemon crate.
-#[allow(dead_code)]
-mod shared_sessions {
-    use std::collections::{BTreeMap, HashMap};
-    use std::path::{Path, PathBuf};
-
-    use serde::{Deserialize, Serialize};
-
-    use crate::app_paths;
-    use crate::engine::EngineType;
-    use crate::shared_binding_visibility::collect_v2_shared_binding_ids_by_session;
-
-    const SHARED_SESSIONS_DIRNAME: &str = "shared-sessions";
-
-    #[derive(Debug, Clone, Serialize)]
-    #[serde(rename_all = "camelCase")]
-    pub(crate) struct SharedSessionSummary {
-        pub(crate) id: String,
-        pub(crate) thread_id: String,
-        pub(crate) title: String,
-        pub(crate) created_at: u64,
-        pub(crate) updated_at: u64,
-        pub(crate) selected_engine: EngineType,
-        pub(crate) thread_kind: String,
-        pub(crate) engine_source: EngineType,
-        pub(crate) selected_engine_label: String,
-        pub(crate) native_thread_ids: Vec<String>,
-    }
-
-    #[derive(Debug, Clone, Deserialize)]
-    #[serde(rename_all = "camelCase")]
-    struct SharedEngineBindingLite {
-        native_thread_id: String,
-    }
-
-    #[derive(Debug, Clone, Deserialize)]
-    #[serde(rename_all = "camelCase")]
-    struct SharedTargetBindingLite {
-        #[serde(default)]
-        native_thread_id: String,
-    }
-
-    #[derive(Debug, Clone, Deserialize)]
-    #[serde(rename_all = "camelCase")]
-    struct SharedSessionMetaLite {
-        id: String,
-        title: String,
-        #[serde(default)]
-        created_at: u64,
-        updated_at: u64,
-        selected_engine: EngineType,
-        #[serde(default)]
-        bindings_by_engine: HashMap<EngineType, SharedEngineBindingLite>,
-        #[serde(default)]
-        bindings_by_target: HashMap<String, SharedTargetBindingLite>,
-    }
-
-    fn shared_sessions_root_dir() -> Result<PathBuf, String> {
-        Ok(app_paths::app_home_dir()?.join(SHARED_SESSIONS_DIRNAME))
-    }
-
-    fn workspace_shared_sessions_dir(workspace_id: &str) -> Result<PathBuf, String> {
-        Ok(shared_sessions_root_dir()?.join(workspace_id))
-    }
-
-    fn shared_session_dir(workspace_id: &str, shared_session_id: &str) -> Result<PathBuf, String> {
-        Ok(workspace_shared_sessions_dir(workspace_id)?.join(shared_session_id))
-    }
-
-    fn shared_session_meta_path(
-        workspace_id: &str,
-        shared_session_id: &str,
-    ) -> Result<PathBuf, String> {
-        Ok(shared_session_dir(workspace_id, shared_session_id)?.join("meta.json"))
-    }
-
-    fn is_safe_shared_session_storage_id(value: &str) -> bool {
-        !value.is_empty()
-            && value != "."
-            && value != ".."
-            && value
-                .chars()
-                .all(|ch| ch.is_ascii_alphanumeric() || ch == '-' || ch == '_')
-    }
-
-    fn parse_shared_session_id(thread_id: &str) -> Result<String, String> {
-        let normalized = thread_id.trim();
-        if let Some(rest) = normalized.strip_prefix("shared:") {
-            let shared_session_id = rest.trim();
-            if is_safe_shared_session_storage_id(shared_session_id) {
-                return Ok(shared_session_id.to_string());
-            }
-        }
-        Err(format!("Invalid shared session thread id: {thread_id}"))
-    }
-
-    fn shared_thread_id(shared_session_id: &str) -> String {
-        format!("shared:{shared_session_id}")
-    }
-
-    fn read_shared_session_meta_lite(
-        workspace_id: &str,
-        shared_session_id: &str,
-    ) -> Result<SharedSessionMetaLite, String> {
-        let path = shared_session_meta_path(workspace_id, shared_session_id)?;
-        let raw = std::fs::read_to_string(&path).map_err(|error| error.to_string())?;
-        serde_json::from_str(&raw).map_err(|error| error.to_string())
-    }
-
-    /// 占位 desktop app 的 `SharedEventWriter` 参数槽。
-    /// daemon catalog 调用点始终传 `None`，并使用第三个参数提供的只读 V2
-    /// binding projection。
-    pub(crate) struct SharedEventWriter;
-
-    /// 保持 desktop signature，使 Shared Session Management 调用点 source-compatible，
-    /// 同时 daemon catalog 保持只读。
-    pub(crate) fn list_workspace_shared_sessions(
-        workspace_id: &str,
-        event_writer: Option<&SharedEventWriter>,
-        event_log_path: Option<&Path>,
-    ) -> Result<Vec<SharedSessionSummary>, String> {
-        let directory = workspace_shared_sessions_dir(workspace_id)?;
-        if !directory.exists() {
-            return Ok(Vec::new());
-        }
-        let mut session_metas = Vec::new();
-        for entry in std::fs::read_dir(&directory).map_err(|error| error.to_string())? {
-            let entry = entry.map_err(|error| error.to_string())?;
-            let file_type = entry.file_type().map_err(|error| error.to_string())?;
-            if !file_type.is_dir() {
-                continue;
-            }
-            let shared_session_id = entry.file_name().to_string_lossy().to_string();
-            let meta = match read_shared_session_meta_lite(workspace_id, &shared_session_id) {
-                Ok(meta) => meta,
-                Err(_) => continue,
-            };
-            session_metas.push(meta);
-        }
-
-        let v2_native_thread_ids_by_session = if event_writer.is_none() {
-            let shared_session_ids = session_metas
-                .iter()
-                .map(|meta| meta.id.clone())
-                .collect::<Vec<_>>();
-            match event_log_path {
-                Some(path) => {
-                    match collect_v2_shared_binding_ids_by_session(path, &shared_session_ids) {
-                        Ok(ids_by_session) => ids_by_session,
-                        Err(error) => {
-                            log::warn!(
-                                "[cc_gui_daemon.shared_sessions] read-only V2 binding recovery failed for workspace {}: {}",
-                                workspace_id,
-                                error
-                            );
-                            BTreeMap::new()
-                        }
-                    }
-                }
-                None => BTreeMap::new(),
-            }
-        } else {
-            BTreeMap::new()
-        };
-
-        let mut summaries = Vec::with_capacity(session_metas.len());
-        for meta in session_metas {
-            let mut native_thread_ids = meta
-                .bindings_by_engine
-                .values()
-                .map(|binding| binding.native_thread_id.clone())
-                .collect::<Vec<_>>();
-            native_thread_ids.extend(
-                meta.bindings_by_target
-                    .values()
-                    .map(|binding| binding.native_thread_id.clone())
-                    .filter(|native_thread_id| !native_thread_id.trim().is_empty()),
-            );
-            if let Some(v2_native_thread_ids) = v2_native_thread_ids_by_session.get(&meta.id) {
-                native_thread_ids.extend(v2_native_thread_ids.iter().cloned());
-            }
-            native_thread_ids.sort();
-            native_thread_ids.dedup();
-            summaries.push(SharedSessionSummary {
-                id: meta.id.clone(),
-                thread_id: shared_thread_id(&meta.id),
-                title: meta.title.clone(),
-                created_at: if meta.created_at > 0 {
-                    meta.created_at
-                } else {
-                    meta.updated_at
-                },
-                updated_at: meta.updated_at,
-                selected_engine: meta.selected_engine,
-                thread_kind: "shared".to_string(),
-                engine_source: meta.selected_engine,
-                selected_engine_label: meta.selected_engine.display_name().to_string(),
-                native_thread_ids,
-            });
-        }
-        summaries.sort_by(|left, right| right.updated_at.cmp(&left.updated_at));
-        Ok(summaries)
-    }
-
-    /// Returns `Ok(true)` when files were removed, `Ok(false)` when already absent.
-    pub(crate) fn delete_shared_session_files(
-        workspace_id: &str,
-        thread_id: &str,
-    ) -> Result<bool, String> {
-        let shared_session_id = parse_shared_session_id(thread_id)?;
-        let path = shared_session_dir(workspace_id, &shared_session_id)?;
-        if !path.exists() {
-            return Ok(false);
-        }
-        std::fs::remove_dir_all(&path).map_err(|error| error.to_string())?;
-        Ok(true)
-    }
-}
+#[path = "cc_gui_daemon/shared_sessions.rs"]
+mod shared_sessions;
 #[allow(dead_code)]
 #[path = "../session_management.rs"]
 mod session_management;
@@ -472,160 +107,11 @@ mod workspace_io;
 #[path = "../workspaces/settings.rs"]
 mod workspace_settings;
 
-// Provide feature-style module paths for shared cores when compiled in the daemon.
-mod codex {
-    pub(crate) type WorkspaceSession = crate::backend::app_server::WorkspaceSession;
-    pub(crate) use crate::codex_doctor::{
-        dsh_node_requirement_error, node_satisfies_dsh_requirement,
-        run_claude_doctor_with_settings, run_codex_doctor_with_settings,
-        run_dsh_doctor_with_settings, run_grok_doctor_with_settings, run_kimi_doctor_with_settings,
-        run_opencode_doctor_with_settings, run_pi_doctor_with_settings,
-        run_qoder_doctor_for_profile_with_settings, run_qoder_doctor_with_settings,
-    };
-    pub(crate) use crate::codex_installer::{
-        build_cli_install_plan_with_backend, resolve_cli_version_status,
-        run_cli_installer_with_progress, CliInstallBackend, CliInstallProgressEvent,
-    };
-    pub(crate) async fn ensure_codex_session(
-        _workspace_id: &str,
-        _state: &crate::state::AppState,
-        _app: &tauri::AppHandle,
-    ) -> Result<(), String> {
-        Err("runtime control commands are unavailable in daemon mode".to_string())
-    }
-    pub(crate) mod args {
-        pub(crate) use crate::codex_args::*;
-    }
-    pub(crate) mod config {
-        pub(crate) use crate::codex_config::*;
-    }
-    pub(crate) mod home {
-        pub(crate) use crate::codex_home::*;
-    }
-    pub(crate) mod launch_profile {
-        pub(crate) use crate::codex_launch_profile::*;
-    }
-    pub(crate) mod provider_env {
-        pub(crate) use crate::codex_provider_env::*;
-    }
-    pub(crate) mod provider_profile {
-        use crate::session_management::CodexProviderBinding;
-        use crate::types::CodexCustomModel;
+#[path = "cc_gui_daemon/codex.rs"]
+mod codex;
 
-        pub(crate) const CODEX_DISK_PROVIDER_PROFILE_ID: &str = "__disk__";
-        pub(crate) const CODEX_DISK_PROVIDER_PROFILE_NAME: &str = "codex-tui/default-config";
-
-        pub(crate) fn codex_provider_binding_for_profile_id(
-            provider_profile_id: &str,
-        ) -> CodexProviderBinding {
-            let provider_profile_id = provider_profile_id.trim();
-            if provider_profile_id.is_empty()
-                || provider_profile_id == CODEX_DISK_PROVIDER_PROFILE_ID
-            {
-                return CodexProviderBinding {
-                    provider_profile_id: CODEX_DISK_PROVIDER_PROFILE_ID.to_string(),
-                    provider_profile_source: "disk".to_string(),
-                    provider_profile_name: CODEX_DISK_PROVIDER_PROFILE_NAME.to_string(),
-                    provider_availability: "available".to_string(),
-                };
-            }
-            CodexProviderBinding {
-                provider_profile_id: provider_profile_id.to_string(),
-                provider_profile_source: "managed".to_string(),
-                provider_profile_name: provider_profile_id.to_string(),
-                provider_availability: "unavailable".to_string(),
-            }
-        }
-
-        pub(crate) fn codex_runtime_key(workspace_id: &str, provider_profile_id: &str) -> String {
-            let provider_profile_id = provider_profile_id.trim();
-            let provider_profile_id = if provider_profile_id.is_empty() {
-                CODEX_DISK_PROVIDER_PROFILE_ID
-            } else {
-                provider_profile_id
-            };
-            format!("codex::{workspace_id}::{provider_profile_id}")
-        }
-
-        pub(crate) fn legacy_codex_runtime_key(workspace_id: &str) -> String {
-            workspace_id.to_string()
-        }
-
-        pub(crate) fn resolve_codex_provider_model_config(
-            provider_profile_id: &str,
-        ) -> Result<Option<(String, Vec<CodexCustomModel>)>, String> {
-            let provider_profile_id = provider_profile_id.trim();
-            if provider_profile_id.is_empty()
-                || provider_profile_id == CODEX_DISK_PROVIDER_PROFILE_ID
-            {
-                return Ok(None);
-            }
-            let path = crate::app_paths::config_file_path()?;
-            let content = std::fs::read_to_string(&path).map_err(|error| {
-                format!("failed to read provider config {}: {error}", path.display())
-            })?;
-            let config: serde_json::Value = serde_json::from_str(&content).map_err(|error| {
-                format!(
-                    "failed to parse provider config {}: {error}",
-                    path.display()
-                )
-            })?;
-            let provider = config
-                .get("codex")
-                .and_then(|codex| codex.get("providers"))
-                .and_then(|providers| providers.get(provider_profile_id))
-                .ok_or_else(|| format!("Codex provider {provider_profile_id} not found"))?;
-            let provider_name = provider
-                .get("name")
-                .and_then(serde_json::Value::as_str)
-                .map(str::trim)
-                .filter(|value| !value.is_empty())
-                .unwrap_or(provider_profile_id);
-            let config_toml = provider
-                .get("configToml")
-                .and_then(serde_json::Value::as_str)
-                .unwrap_or_default()
-                .trim()
-                .to_string();
-            if config_toml.is_empty() {
-                return Err(format!(
-                    "Codex provider {provider_name} has empty configToml"
-                ));
-            }
-            let custom_models = provider
-                .get("customModels")
-                .cloned()
-                .map(serde_json::from_value)
-                .transpose()
-                .map_err(|error| {
-                    format!("Codex provider {provider_name} has invalid customModels: {error}")
-                })?
-                .unwrap_or_default();
-            Ok(Some((config_toml, custom_models)))
-        }
-    }
-    pub(crate) mod rewind {
-        pub(crate) use crate::codex_rewind::*;
-    }
-    pub(crate) mod collaboration_policy {
-        pub(crate) use crate::codex_collaboration_policy::*;
-    }
-    pub(crate) mod thread_mode_state {
-        pub(crate) use crate::codex_thread_mode_state::*;
-    }
-}
-
-mod files {
-    pub(crate) mod io {
-        pub(crate) use crate::file_io::*;
-    }
-    pub(crate) mod ops {
-        pub(crate) use crate::file_ops::*;
-    }
-    pub(crate) mod policy {
-        pub(crate) use crate::file_policy::*;
-    }
-}
+#[path = "cc_gui_daemon/files.rs"]
+mod files;
 
 use serde::{Deserialize, Serialize};
 use serde_json::{json, Value};
