+++ src-tauri/src/bin/cc_gui_daemon/files.rs
@@ -1,0 +1,10 @@
+pub(crate) mod io {
+    pub(crate) use crate::file_io::*;
+}
+pub(crate) mod ops {
+    pub(crate) use crate::file_ops::*;
+}
+pub(crate) mod policy {
+    pub(crate) use crate::file_policy::*;
+}
+
